feature: extract constants
