+++ src/main/kotlin/net/atos/zac/mail/MailMessageBuilder.kt
@@ -26,24 +26,6 @@ class MailMessageBuilder(
     private var body: String,
     private var attachments: List<Attachment>
 ) {
-    private fun setupPasswordAuthentication(mailSession: Session) {
-        val userName = System.getenv("SMTP_USERNAME")
-        val password = System.getenv("SMTP_PASSWORD")
-        userName?.let {
-            mailSession.properties.setProperty("mail.smtp.user", userName)
-            mailSession.properties.setProperty("mail.smtp.auth", "true")
-        }
-        password?.let {
-            val authSection = if (userName != null) "$userName@" else ""
-            val smtpServerHost = System.getenv("SMTP_SERVER")
-            mailSession.setPasswordAuthentication(
-                URLName("smtp://$authSection$smtpServerHost"),
-                PasswordAuthentication(userName, password)
-            )
-            mailSession.properties.setProperty("mail.smtp.auth", "true")
-        }
-    }
-
     fun build(mailSession: Session) = setupPasswordAuthentication(mailSession).let {
         MimeMessage(mailSession).apply {
             addHeader("Content-type", "${MediaType.TEXT_HTML}; charset=${StandardCharsets.UTF_8.name()}")
